requirements.txt
fpdf==1.7.2
oracledb==3.4.2
pandas==2.3.3
Pillow==11.3.0
pdfplumber
python-docx
docx2pdf
google-generativeai
python-dotenv
